test: add test module for fusion security with custom context vaadin/spring#812
IT tests for vaadin/flow#10997

It reuses the code in the test-spring-security-fusion module. It also packages the frontend folder in test-spring-security-fusion as a jar and extracts the jar to the new test module as the frontend folder.

It has to copy the Endpoint files, as the endpoint generation can only detect the files in the source.