fix: Do not sort class fields (#1180)
The natural order is... natural. This allows using the TS model to generate e.g. grid columns in the same order the properties are in the bean

Do note that some JDKs do not always provide all methods and fields in the declaration order. The order used will be whatever the JDK returns